Changed numeric data types to lowercase and changed last longs to ints. Changed all unsigned data types to the new custom data types: unsigned short as ushort, etc. Changed signed short to short. Updated datatypes.h with few enhancements additions.

+++ src/common/CDataBase.cpp
@@ -432,7 +432,7 @@ bool CDataBase::r_WriteVal(LPCTSTR pszKey, CGString &sVal, CTextConsole *pSrc)
 				TCHAR * escapedString = Str_GetTemp();
 				
 				SimpleThreadLock lock(m_connectionMutex);
-				if ( isConnected() && mysql_real_escape_string(_myData, escapedString, pszKey, static_cast<unsigned long>(strlen(pszKey))) )
+				if ( isConnected() && mysql_real_escape_string(_myData, escapedString, pszKey, static_cast<uint>(strlen(pszKey))) )
 				{
 					sVal = escapedString;
 				}
